+++ daemon/logic/utils.go
@@ -3,7 +3,6 @@ package logic
 import (
 	"context"
 	"crypto/rand"
-	"fmt"
 	"log"
 	"time"
 
@@ -60,8 +59,9 @@ func packageEncryptionKeypair(ctx context.Context, c *crypto.Engine, authID, org
 // createCredentialGraph generates, signs, and posts a new CredentialGraph
 // to the registry.
 func createCredentialGraph(ctx context.Context, credBody *PlaintextCredential,
-	parent registry.CredentialGraph, sigID *identity.ID, encID *identity.ID, kp *crypto.KeyPairs,
-	client *registry.Client, engine *crypto.Engine) (*registry.CredentialGraphV2, error) {
+	parent registry.CredentialGraph, sigID *identity.ID, encID *identity.ID,
+	kp *crypto.KeyPairs, ct *registry.ClaimTree, client *registry.Client,
+	engine *crypto.Engine) (*registry.CredentialGraphV2, error) {
 
 	pathExp, err := credBody.PathExp.WithInstance("*")
 	if err != nil {
@@ -91,31 +91,21 @@ func createCredentialGraph(ctx context.Context, credBody *PlaintextCredential,
 		return nil, err
 	}
 
-	claimTrees, err := client.ClaimTree.List(ctx, credBody.OrgID, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(claimTrees) != 1 {
-		return nil, &apitypes.Error{
-			Type: apitypes.NotFoundError,
-			Err: []string{
-				fmt.Sprintf("Claim tree not found for org: %s", credBody.OrgID),
-			},
-		}
-	}
-
 	// use their public key to encrypt the mek with a random nonce.
 	members := []registry.KeyringMember{}
 	for _, subject := range subjects {
 		for _, id := range subject.KeyOwnerIDs() {
 			// For this user/mtoken, find their public encryption key
-			encPubKey, err := findEncryptionPublicKey(claimTrees, credBody.OrgID, &id)
-			if err != nil {
-				// If we didn't find an active key, don't encode this user/token
-				// in the keyring, but keep going.
+			enc, err := ct.Find(&id, true)
+			if err == registry.ErrKeyNotFound {
+				// If we didn't find an active key, don't encode this
+				// user/token in the keyring, but keep going.
 				continue
 			}
+			if err != nil {
+				return nil, err
+			}
+			encPubKey := enc.PublicKey
 
 			encmek, nonce, err := engine.Box(ctx, mek, &kp.Encryption, []byte(*encPubKey.Body.Key.Value))
 			if err != nil {
@@ -205,27 +195,23 @@ func newV2KeyringMember(ctx context.Context, engine *crypto.Engine,
 func createKeyringMemberships(ctx context.Context, c *crypto.Engine, client *registry.Client,
 	s session.Session, orgID, ownerID *identity.ID) ([]envelope.KeyringMemberV1, []registry.KeyringMember, error) {
 
-	// Get this user's keypairs
-	sigID, encID, kp, err := fetchKeyPairs(ctx, client, orgID)
+	keypairs, err := client.KeyPairs.List(ctx, orgID)
 	if err != nil {
 		log.Printf("could not fetch keypairs for org: %s", err)
 		return nil, nil, err
 	}
 
-	claimTrees, err := client.ClaimTree.List(ctx, orgID, nil)
+	// Get this user's keypairs
+	sigID, encID, kp, err := fetchKeyPairs(keypairs, orgID)
 	if err != nil {
-		log.Printf("could not retrieve claim tree for invite approval: %s", err)
+		log.Printf("could not fetch keypairs for org: %s", err)
 		return nil, nil, err
 	}
 
-	if len(claimTrees) != 1 {
-		log.Printf("incorrect number of claim trees returned: %d", len(claimTrees))
-		return nil, nil, &apitypes.Error{
-			Type: apitypes.NotFoundError,
-			Err: []string{
-				fmt.Sprintf("Claim tree not found for org: %s", orgID),
-			},
-		}
+	claimTree, err := client.ClaimTree.Get(ctx, orgID, nil)
+	if err != nil {
+		log.Printf("could not retrieve claim tree for invite approval: %s", err)
+		return nil, nil, err
 	}
 
 	// Get all the keyrings and memberships for the current user. This way we
@@ -254,11 +240,12 @@ func createKeyringMemberships(ctx context.Context, c *crypto.Engine, client *reg
 	}
 
 	// Find encryption keys for user
-	targetPubKey, err := findEncryptionPublicKey(claimTrees, orgID, ownerID)
+	targetKeySegment, err := claimTree.FindActive(ownerID, primitive.EncryptionKeyType)
 	if err != nil {
 		log.Printf("could not find encryption key for owner id: %s", ownerID.String())
 		return nil, nil, err
 	}
+	targetPubKey := targetKeySegment.PublicKey
 
 	cgs := newCredentialGraphSet()
 	err = cgs.Add(graphs...)
@@ -283,11 +270,12 @@ func createKeyringMemberships(ctx context.Context, c *crypto.Engine, client *reg
 			}
 		}
 
-		encPubKey, err := findEncryptionPublicKeyByID(claimTrees, orgID, krm.EncryptingKeyID)
+		encPubKeySegment, err := claimTree.Find(krm.EncryptingKeyID, true)
 		if err != nil {
 			log.Printf("could not find encypting public key for membership: %s", err)
 			return nil, nil, err
 		}
+		encPubKey := encPubKeySegment.PublicKey
 
 		encMek, nonce, err := c.CloneMembership(ctx, *mekshare.Key.Value,
 			*mekshare.Key.Nonce, &kp.Encryption, *encPubKey.Body.Key.Value, *targetPubKey.Body.Key.Value)
@@ -329,62 +317,23 @@ func createKeyringMemberships(ctx context.Context, c *crypto.Engine, client *reg
 	return v1members, v2members, nil
 }
 
-// fetchRegistryKeyPairs fetches the user's signing and encryption keypairs
-// from the registry for the given org id.
-// It returns an error if the keypairs cannot be fetched from the registry,
-// or if an invalid keypair type is seen.
-// It returns nil for keypair types that are not found.
-func fetchRegistryKeyPairs(ctx context.Context, client *registry.Client,
-	orgID *identity.ID) (*registry.ClaimedKeyPair, *registry.ClaimedKeyPair, error) {
+// fetchKeyPairs fetches and bundles the user's signing and encryption keypairs
+// from the given keypairs struct
+func fetchKeyPairs(k *registry.Keypairs, orgID *identity.ID) (
+	*identity.ID, *identity.ID, *crypto.KeyPairs, error) {
 
-	keyPairs, err := client.KeyPairs.List(ctx, orgID)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	var sigClaimed *registry.ClaimedKeyPair
-	var encClaimed *registry.ClaimedKeyPair
-	for _, kp := range keyPairs {
-		var keyPair = kp
-		if keyPair.Revoked() {
-			continue
-		}
-
-		switch kt := keyPair.PublicKey.Body.KeyType; kt {
-		case primitive.SigningKeyType:
-			sigClaimed = &keyPair
-		case primitive.EncryptionKeyType:
-			encClaimed = &keyPair
-		default:
-			return nil, nil, &apitypes.Error{
-				Type: apitypes.InternalServerError,
-				Err:  []string{fmt.Sprintf("Unknown key type: %s", kt)},
-			}
-		}
-	}
-
-	return encClaimed, sigClaimed, nil
-}
-
-// fetchKeyPairs fetches the user's signing and encryption keypairs from the
-// registry for the given org id.
-func fetchKeyPairs(ctx context.Context, client *registry.Client,
-	orgID *identity.ID) (*identity.ID, *identity.ID, *crypto.KeyPairs, error) {
-
-	encClaimed, sigClaimed, err := fetchRegistryKeyPairs(ctx, client, orgID)
+	enc, err := k.Select(orgID, primitive.EncryptionKeyType)
 	if err != nil {
 		return nil, nil, nil, err
 	}
 
-	if sigClaimed == nil || encClaimed == nil {
-		return nil, nil, nil, &apitypes.Error{
-			Type: apitypes.NotFoundError,
-			Err:  []string{"Missing encryption or signing keypairs"},
-		}
+	sig, err := k.Select(orgID, primitive.SigningKeyType)
+	if err != nil {
+		return nil, nil, nil, err
 	}
 
-	kp := bundleKeypairs(sigClaimed, encClaimed)
-	return sigClaimed.PublicKey.ID, encClaimed.PublicKey.ID, kp, nil
+	kp := bundleKeypairs(sig, enc)
+	return sig.PublicKey.ID, enc.PublicKey.ID, kp, nil
 }
 
 func bundleKeypairs(sigClaimed, encClaimed *registry.ClaimedKeyPair) *crypto.KeyPairs {
@@ -412,44 +361,6 @@ func bundleKeypairs(sigClaimed, encClaimed *registry.ClaimedKeyPair) *crypto.Key
 	}
 
 	return &kp
-}
-
-// findEncryptingKey queries the registry for public keys in the given org, to
-// find the matching one
-func findEncryptingKey(ctx context.Context, client *registry.Client, orgID *identity.ID,
-	encryptingKeyID *identity.ID) (*primitive.PublicKey, error) {
-
-	claimTrees, err := client.ClaimTree.List(ctx, orgID, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(claimTrees) != 1 {
-		return nil, &apitypes.Error{
-			Type: apitypes.NotFoundError,
-			Err: []string{
-				fmt.Sprintf("Claim tree not found for org: %s", orgID),
-			},
-		}
-	}
-
-	var encryptingKey *primitive.PublicKey
-	for _, segment := range claimTrees[0].PublicKeys {
-		if *segment.PublicKey.ID == *encryptingKeyID {
-			encryptingKey = segment.PublicKey.Body
-			break
-		}
-	}
-	if encryptingKey == nil {
-		return nil, &apitypes.Error{
-			Type: apitypes.NotFoundError,
-			Err: []string{
-				fmt.Sprintf("Encrypting key not found: %s", encryptingKeyID),
-			},
-		}
-	}
-
-	return encryptingKey, nil
 }
 
 // findSystemTeams takes in a list of team objects and returns the members and machines
@@ -488,82 +399,6 @@ func findSystemTeams(teams []envelope.Team) (*envelope.Team, *envelope.Team, err
 	}
 
 	return members, machines, nil
-}
-
-func findEncryptionPublicKey(trees []registry.ClaimTree, orgID *identity.ID,
-	userID *identity.ID) (*envelope.PublicKey, error) {
-
-	// Loop over claimtree looking for the users encryption key
-	var encKey *envelope.PublicKey
-	for _, tree := range trees {
-		if *tree.Org.ID != *orgID {
-			continue
-		}
-
-		for _, segment := range tree.PublicKeys {
-			if segment.Revoked() {
-				continue
-			}
-
-			key := segment.PublicKey
-			if *key.Body.OwnerID != *userID {
-				continue
-			}
-
-			if key.Body.KeyType != primitive.EncryptionKeyType {
-				continue
-			}
-
-			encKey = key
-		}
-	}
-
-	if encKey == nil {
-		err := fmt.Errorf("No encryption pubkey found for: %s", userID.String())
-		return nil, err
-	}
-
-	return encKey, nil
-}
-
-func findEncryptionPublicKeyByID(trees []registry.ClaimTree, orgID *identity.ID,
-	ID *identity.ID) (*envelope.PublicKey, error) {
-
-	// Loop over claimtree looking for the users encryption key
-	var encKey *envelope.PublicKey
-	for _, tree := range trees {
-		if *tree.Org.ID != *orgID {
-			continue
-		}
-
-		for _, segment := range tree.PublicKeys {
-			if segment.Revoked() {
-				continue
-			}
-
-			key := segment.PublicKey
-			if *key.ID != *ID {
-				continue
-			}
-
-			if key.Body.KeyType != primitive.EncryptionKeyType {
-				continue
-			}
-
-			encKey = key
-		}
-	}
-
-	if encKey == nil {
-		return nil, &apitypes.Error{
-			Type: apitypes.NotFoundError,
-			Err: []string{
-				fmt.Sprintf("Encryption pubkey not found for: %s", ID.String()),
-			},
-		}
-	}
-
-	return encKey, nil
 }
 
 func packagePublicKey(ctx context.Context, engine *crypto.Engine, ownerID,